Fix deadlock caused by JS callback not being called when scanner stopped/paused from the JS side (#201) (#202)
It's possible that we are requested to stop/pause the picker while we are waiting for a callback from the JS side. When the JS side asks to stop/pause the picker, this callback is not executed and the signal is not sent to the semaphore. This means that the session queue, that is waiting for the semaphore, will just wait forever. Here by signaling the semaphores, we make sure that the session queue is not blocked and it will finish the code it has to execute.

Signed-off-by: Luca Torella <luca@scandit.com>
(cherry picked from commit 0216f0b1ad53e899f9e5c5c00985df5035c4a423)